+++ online-bookstore/online-bookstore-domain/src/main/java/es/pedrazamiguez/assessment/onlinebookstore/domain/usecase/inventory/AddToInventoryUseCase.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 @FunctionalInterface
 public interface AddToInventoryUseCase {
 
-  Optional<BookAllocation> addToInventory(Long bookId, int copies);
+  Optional<BookAllocation> addToInventory(Long bookId, Long copies);
 }
